Refactor availability button
