+++ lxl-web/src/routes/(app)/[[lang=lang]]/[fnurgel=fnurgel]/HoldingInfo.svelte
@@ -1,33 +1,34 @@
 <script lang="ts">
 	import { page } from '$app/state';
-	import type { DecoratedHolder, ItemLinksByBibId } from '$lib/types/holdings';
-	import { ShowLabelsOptions } from '$lib/types/decoratedData';
-	import DecoratedData from '$lib/components/DecoratedData.svelte';
-	import BiChevronRight from '~icons/bi/chevron-right';
+	import type { BibIdObj, DecoratedHolder, ItemLinksByBibId } from '$lib/types/holdings';
 	import { BibDb } from '$lib/types/xl';
-	import HoldingAvailability from './HoldingAvailability.svelte';
+	import type { HoldingStatus } from '$lib/types/api';
+	import BiChevronRight from '~icons/bi/chevron-right';
+	import BiChevronDown from '~icons/bi/chevron-down';
 
 	type HoldingInfoProps = {
+		id: BibIdObj;
 		holder: DecoratedHolder;
-		holdingUrl: string;
 		linksByBibId: ItemLinksByBibId;
+		isSingle: boolean;
 	};
 
-	const { holder, holdingUrl, linksByBibId }: HoldingInfoProps = $props();
+	const { id, holder, linksByBibId, isSingle }: HoldingInfoProps = $props();
 
 	const sigel = holder?.sigel;
-	let instancesCollapsed = $state(true);
+	let loading = $state(false);
+	let expandedHoldingAvailability = $state(false);
+	let statusData: HoldingStatus[] | undefined = $state();
+	let error: string | undefined = $state();
+	let bibIds: BibIdObj[] = [id];
 
-	// if holdingUrl is an instance fnurgel, add its mapped bibId object into arr,
-	// else add all ids of current type with holdings for current sigel
-	const bibIds = $derived.by(() => {
-		return page.data.bibIdsByInstanceId?.[holdingUrl]
-			? [page.data.bibIdsByInstanceId[holdingUrl]]
-			: Object.keys(page.data.bibIdsByInstanceId)
-					.filter((i) => page.data.bibIdsByInstanceId[i]['@type'] === holdingUrl)
-					.filter((i) => page.data.bibIdsByInstanceId[i].holders?.includes(sigel))
-					.map((i) => page.data.bibIdsByInstanceId[i]);
-	});
+	function hasInstanceContent(id: string) {
+		return hasLoanReserveLink(id) || hasLinksToItem(id) || hasStatusLink(id);
+	}
+
+	function hasStatusLink(id: string) {
+		return linksByBibId[id][holder.sigel][BibDb.ItemStatus];
+	}
 
 	function hasLinksToItem(id: string) {
 		return linksByBibId[id]?.[holder.sigel]?.[BibDb.LinksToItem];
@@ -37,126 +38,207 @@
 		return linksByBibId[id]?.[holder.sigel]?.[BibDb.LoanReserveLink];
 	}
 
-	function missingAtLeastOneLinkToItem() {
-		return bibIds.find((id) => !hasLinksToItem(id.bibId) && !hasLoanReserveLink(id.bibId));
+	async function holdingStatusClicked() {
+		expandedHoldingAvailability = !expandedHoldingAvailability;
+		await getHoldingStatus();
+	}
+	async function getHoldingStatus() {
+		if (!sigel || !bibIds || bibIds.length < 1) {
+			error = page.data.t('holdings.loanStatusFailed');
+		} else if (!statusData) {
+			loading = true;
+
+			fetchHoldingStatus(bibIds)
+				.then((data) => {
+					statusData = data;
+					loading = false;
+				})
+				.catch((err) => {
+					error = `${page.data.t('holdings.loanStatusFailed')}: ${err}`;
+					loading = false;
+				});
+		}
 	}
 
-	function hasStatusLink(id: string) {
-		return linksByBibId[id][holder.sigel][BibDb.ItemStatus];
+	function urlNotDefinedError(err: string | undefined) {
+		if (
+			err &&
+			err === `URL till lånestatus för ${sigel} ej definierad i LIBRIS biblioteksdatabas!`
+		) {
+			return true;
+		}
+		return false;
 	}
 
-	function hasInstanceContent(id: string) {
-		return hasLoanReserveLink(id) || hasLinksToItem(id) || hasStatusLink(id);
+	async function fetchHoldingStatus(ids: BibIdObj[]) {
+		const promises = ids.map((id) => {
+			if (id) {
+				const searchParams = new URLSearchParams();
+				searchParams.set('bib_id', id.bibId);
+				searchParams.set('sigel', sigel);
+				if (id.onr) {
+					searchParams.set('onr', id.onr);
+				}
+				id.isbn.forEach((i: string) => {
+					searchParams.set('isbn', i);
+				});
+				id.issn.forEach((i: string) => {
+					searchParams.set('issn', i);
+				});
+				return fetch(`/api/holdingstatus?${searchParams.toString()}`);
+			}
+		});
+
+		const responses = await Promise.all(promises);
+		const data = await Promise.all(
+			responses.map((response) => {
+				if (response && response.ok) {
+					return response.json();
+				} else {
+					error = `${page.data.t('holdings.loanStatusFailed')}: ${response?.status} ${response?.statusText}`;
+				}
+			})
+		);
+		return data;
+	}
+
+	function getIndicator(status: string) {
+		if (status && typeof status === 'string') {
+			switch (status.toLowerCase()) {
+				case 'tillgänglig':
+				case 'ej utlånad':
+				case 'available':
+					return 'available';
+				case 'ej tillgänglig':
+				case 'utlånad':
+				case 'on loan':
+				case 'not available':
+					return 'unavailable';
+				default:
+					return false;
+			}
+		}
+		return false;
 	}
 </script>
 
-<li class="border-neutral text-sm not-last:border-b">
-	<h2 class="holder-label mt-4">
-		<DecoratedData data={holder.obj} showLabels={ShowLabelsOptions['Never']} />
-	</h2>
-	<div class="">
-		<ul class="bib-list" class:instancesCollapsed>
-			{#each bibIds as id (id.bibId)}
-				{#if linksByBibId[id.bibId]?.[holder.sigel]}
-					{@const linksForHolder = linksByBibId[id.bibId][holder.sigel]}
-					<li>
-						{#if hasInstanceContent(id.bibId)}
-							{#if bibIds.length > 1}
-								<div class="text-subtle">{id.str || '-'}</div>
-							{/if}
-							{#if hasLoanReserveLink(id.bibId)}
-								<a
-									href={linksForHolder['loanReserveLink'].at(0)}
-									target="_blank"
-									class="ext-link mr-2 h-9 max-w-sm align-top"
-								>
-									{page.data.t('holdings.loanReserveLink')}
-								</a>
-							{:else if hasLinksToItem(id.bibId)}
-								<a
-									href={linksForHolder[BibDb.LinksToItem].at(0)}
-									target="_blank"
-									class="ext-link mr-2 h-9 max-w-sm align-top"
-								>
-									{page.data.t('holdings.linkToLocal')}
-								</a>
-							{/if}
-							{#if hasStatusLink(id.bibId)}
-								<HoldingAvailability {holder} bibId={id}></HoldingAvailability>
-							{/if}
-						{/if}
-					</li>
-				{/if}
-			{/each}
-		</ul>
-		<ul>
-			{#if bibIds.length > 5}
-				<li>
-					<button class="show-all" onclick={() => (instancesCollapsed = !instancesCollapsed)}>
-						{#if instancesCollapsed}
-							{page.data.t('holdings.showAll')} ({bibIds.length})
-						{:else}
-							{page.data.t('holdings.showFewer')}...
-						{/if}
-					</button>
-				</li>
-			{/if}
-			{#if bibIds.at(0) && missingAtLeastOneLinkToItem()}
-				{@const firstBibId = bibIds.at(0).bibId}
-				<li>
-					{#if linksByBibId[firstBibId]?.[holder.sigel]?.[BibDb.LinksToCatalog]}
-						<a
-							href={linksByBibId[firstBibId][holder.sigel][BibDb.LinksToCatalog].at(0)}
-							target="_blank"
-							class="ext-link"
-						>
-							{page.data.t('holdings.linkToCatalog')}
-						</a>
-					{:else if linksByBibId[firstBibId]?.[holder.sigel]?.[BibDb.LinksToSite]}
-						<a
-							href={linksByBibId[firstBibId][holder.sigel][BibDb.LinksToSite].at(0)}
-							target="_blank"
-							class="ext-link"
-						>
-							{page.data.t('holdings.linkToSite')}
-						</a>
-					{/if}
-				</li>
-			{/if}
-		</ul>
-	</div>
-	<details>
-		<summary class="my-3 flex cursor-pointer items-baseline">
-			<span
-				class="text-3xs arrow text-subtle mr-0.5 h-3 origin-center rotate-0 transition-transform"
+{#if linksByBibId[id.bibId]?.[holder.sigel]}
+	{@const linksForHolder = linksByBibId[id.bibId][holder.sigel]}
+	{#if hasInstanceContent(id.bibId)}
+		{#if isSingle}
+			<div class="text-subtle">{id.str || '-'}</div>
+		{/if}
+		{#if hasLoanReserveLink(id.bibId)}
+			<a
+				href={linksForHolder['loanReserveLink'].at(0)}
+				target="_blank"
+				class="ext-link mr-2 h-9 max-w-sm align-top"
 			>
-				<BiChevronRight />
-			</span>
-			{page.data.t('holdings.openingHoursEtc')}
-		</summary>
-		<div class="status-container border-neutral bg-page my-3 max-w-md rounded-sm border p-2">
-			<span>
-				{#if bibIds.at(0)}
-					{@const firstBibId = bibIds.at(0).bibId}
-					<ul style="white-space: pre-line">
-						{#if linksByBibId[firstBibId]?.[holder.sigel]?.[BibDb.OpeningHours]}
-							<li>
-								{linksByBibId[firstBibId][holder.sigel][BibDb.OpeningHours].at(0)}
-							</li>
-						{/if}
-						{#if linksByBibId[firstBibId]?.[holder.sigel]?.[BibDb.Address]}
-							{#each linksByBibId[firstBibId][holder.sigel][BibDb.Address] as address, index (index)}
-								<li class="my-2">
-									{address}
-								</li>
-							{/each}
-						{/if}
-					</ul>
+				{page.data.t('holdings.loanReserveLink')}
+			</a>
+		{:else if hasLinksToItem(id.bibId)}
+			<a
+				href={linksForHolder[BibDb.LinksToItem].at(0)}
+				target="_blank"
+				class="ext-link mr-2 h-9 max-w-sm align-top"
+			>
+				{page.data.t('holdings.linkToLocal')}
+			</a>
+		{/if}
+		{#if hasStatusLink(id.bibId)}
+			<button onclick={holdingStatusClicked}>
+				<div class="text-subtle link-subtle flex cursor-pointer items-baseline">
+					{#if !expandedHoldingAvailability}
+						<span
+							class="text-3xs arrow text-subtle mr-0.5 h-3 origin-center rotate-0 transition-transform"
+						>
+							<BiChevronRight />
+						</span>
+					{:else}
+						<span
+							class="text-3xs arrow text-subtle mr-0.5 h-3 origin-center rotate-0 transition-transform"
+						>
+							<BiChevronDown />
+						</span>
+					{/if}
+					{page.data.t('holdings.loanStatus')}
+				</div>
+			</button>
+			<div class="mb-4">
+				{#if loading}
+					<p>{page.data.t('search.loading')}</p>
 				{/if}
-			</span>
-		</div>
-	</details>
-</li>
+				{#if error}
+					<div class="status-container border-neutral bg-page mt-2 max-w-md rounded-sm border p-2">
+						<p class="error" role="alert">{error}</p>
+					</div>
+				{/if}
+				{#if statusData && statusData.length > 0 && expandedHoldingAvailability}
+					{#each statusData as instance, index (index)}
+						{#if instance?.item_information}
+							{@const items = instance.item_information}
+							<div
+								class="status-container border-neutral bg-page mt-2 flex max-w-md flex-col gap-4 rounded-sm border p-2"
+							>
+								{#if items.error || items.count === 0}
+									{#if urlNotDefinedError(items.error)}
+										<p class="library-unavailable">{page.data.t('holdings.libraryUnvaliable')}</p>
+									{:else}
+										<p class="error" role="alert">{page.data.t('holdings.loanStatusFailed')}</p>
+									{/if}
+								{/if}
+								{#each items.items as item, index (index)}
+									{@const indicator = getIndicator(item.Status)}
+									<table>
+										<tbody class="text-xs leading-relaxed">
+											<tr>
+												<th>{page.data.t('holdings.location')}</th>
+												<td>{item.Location}</td>
+											</tr>
+											<tr>
+												<th>{page.data.t('holdings.shelf')}</th>
+												<td>
+													{item.Call_No}
+													<!-- show map link only if absolute url -->
+													{#if item.Map && (item.Map.startsWith('http://') || item.Map.startsWith('https://'))}
+														(<a href={item.Map} target="_blank" class="ext-link">
+															{page.data.t('holdings.map')}</a
+														>)
+													{/if}
+												</td>
+											</tr>
+											<tr>
+												<th>{page.data.t('holdings.loanPolicy')}</th>
+												<td>{item.Loan_Policy}</td>
+											</tr>
+											<tr>
+												<th>{page.data.t('holdings.status')}</th>
+												<td>
+													{#if indicator}
+														<span class={`indicator ${indicator}`}></span>
+														{page.data.t(`holdings.${indicator}`)}
+													{:else}
+														{item.Status}
+													{/if}
+												</td>
+											</tr>
+											{#if item.Status_Date}
+												<tr>
+													<th>{page.data.t('holdings.date')}</th>
+													<td>{item?.Status_Date_Description} {item.Status_Date}</td>
+												</tr>
+											{/if}
+										</tbody>
+									</table>
+								{/each}
+							</div>
+						{/if}
+					{/each}
+				{/if}
+			</div>
+		{/if}
+	{/if}
+{/if}
 
 <style lang="postcss">
 	@reference "../../../../app.css";
